Pull out Syntax field from Entry
Pull out Syntax field from Entry because the proto2 and proto3 spec
say the syntax field must come first and can't be duplicated.

Having the Syntax field separate helps in the parse step to avoid
walking the parse tree twice as we need to know if we are in proto2 or
proto3 syntax at various points during codegen.

https://developers.google.com/protocol-buffers/docs/reference/proto2-spec#proto_file
https://developers.google.com/protocol-buffers/docs/reference/proto3-spec#proto_file