Add reserved fields and extensions ranges
Add reserved fields and extensions ranges to FileDescriptor output.
Extensions for ordinary messages are only valid in proto2, so no proto3
tests can be included. The range calculation and range max for fields
and extensions are identical, but different from Enum range calculation
where the end is excludsive and max is int32 max, rather than 1 << 29 -
this is a protoc implementation detail with unknown cause.