Add oneof declarations
Add oneof declarations to compiler and fileDescriptor. While at it fix a
bug with proto3 optional fields: each optional field is appended to
the _end_ of all oneof declarations and the oneofIndex references this
declaration in the array of oneofDeclarations (and not the field index
as initially assumed). Add tests confirming this behaviour.