Add extensions to FileDescriptors
Add extensions to FileDescriptors, nested within message and on the top
level. Proto3 only supports extending
File|Message|Field|Service|Method|Enum etc. Options, rather than
extending arbitrary messages (the `any` type is provided for that) - so
in order to test extensions with proto3 we have extended

	google.protobuf.MethodOptions

However, the underlying `google/protobuf/descriptor.proto` file cannot
yet be fully translated to a FileDescriptor, as we are still missing
the implementation of options. To work around this we use
the _no_include.proto file suffix and rework the Makefile a little bit
to not include dependencies in the FileDescriptor for proto files
ending in that suffix. For this purpose we have also copied the `06_...`
examples.

Extensions have now been tested with the typical golden tests compared
to protoc FileDescriptor output, but we are still missing `group`
fields for proto2.

Signed-off-by: Julia Ogris <julia.ogris@gmail.com>