Fix Group messages from Extensions
Add Group message definitions from Extensions to top level and Nested
Message definitions in FileDescriptor set. This has been overlooked in
the previous implementation of groups.

In order to make this work properly and be byte-by-byte comparable to
protoc fileDescriptors Groups created in Extensioins must be listed in
the order they appear in the original .proto source file, interspersed
with other message definitions. For this purpose we can no longer
process extensions after processing messages and we therefore follow
the builder pattern used for Messages and Fields, including the
builder.AddEntry method.