Update parser_test.go for valid option input
Update parser_test.go for valid option input, the updated snippet didn't
compile with protoc and caused a little bit of confusion on my end.
No option can start with `.` iiuc. Builtin options cannot be fully qualified.
e.g.

	option deprecated = true;

is valid, but none of the following:

	option google.protobuf.deprecated = true;
	option .google.protobuf.deprecated = true;
	option google.protobuf.MessageOptions.deprecated = true;
	option .google.protobuf.MessageOptions.deprecated = true;

similar custom options, defined as extension _always_ have to start with
`(` - that's how extension fields are identified, e.g.

	option (priority) = 1;
	option (pkg.priority) = 1;
	option (.pkg.priority) = 1;

None of them take a leading `.` dot or a leading package qualifier, so
the following is invalid and doesn't compile:

	option pkg.(priority) = 1;
	option .pkg.(priority) = 1;
	option .(priority) = 1;