Compile MessageOption
Compile MessageOption, by populating fields from
google.protobuf.MessageOption directly and saving extension fields in
the MessageDescriptor as []*UninterpretedOption, with the name broken
up as slice into direct field references and extension field
references.

The goal here is to lay the groundwork to correctly interpret and save
custom options in the FileDescriptor.

In the follow up commits we will try and close the loop for custom
Message Options and then continue on to all other types of options.

Signed-off-by: Julia Ogris <julia.ogris@gmail.com>