Add all builtin, native, non-custom options
Add all "native", non-custom options as they are defined in the
descriptor.proto:

*  FileOptions
*  EnumOptions
*  EnumValueOptions
*  MessageOptions
*  FieldOptions
*  OneofOptions
*  ExtensionRangeOptions
*  ServiceOptions
*  MethodOptions

Extend pariticple parser to parse Extension Options and return Option
Value as string as this is needed in the "DefaultValue" FieldOption.
Add integration/golden tests for all options, except
ExtensionRangeOptions and OneofOptions as these do not have builtin
options and only work with custom options which aren't ready yet.

Signed-off-by: Julia Ogris <julia.ogris@gmail.com>