compiler: Make types map into a struct
Turn the `types` map into a struct so that we can add tracking of
extension names in addition to the types. Extension names can be
referred to before they are defined and they can also be referred to
in a scoped manner like types (i.e. omitting the package name if used in
the same package as defined, and same for nesting within Messages).