parser: Capture option field names separately
Capture all the name parts of an option as separate name elements.
For example:

  option (pkg.foo).bar.baz.(pkg2.Message1.qux) = 5;

has 4 elements: "(pkg.foo)", "bar", "baz" and "(pkg2.Message1.qux)".
Each element is a single name, either a field or extension, that can be
set in a message.

Prior to this change, the parser was capturing "bar.baz" as a single
name element which would need to be further split to look up field
names.