tests: Enable tests that now pass
Enable the conformance tests:
  unittest_custom_options.proto
  unittest_lazy_dependencies.proto
  unittest_proto3_optional.proto

and the compiler test:
  17_proto2_custom_options.proto

as these tests now pass now that options are handled properly.

Alter how protos are compared in the tests so that they pass. When the
golden files were read in to compare to the generated protos, the
extensions were not being resolved, so those extension fields just
appeared as `RawField`s. These do not compare well as the same proto
messages can and do have different underlying representations. By
properly resolving extensions when loading the protobuf files, they are
normalised properly.

The previous way of trying to compare protobuf files by writing out what
the compiler generates and reading it back in has been removed as it is
no longer necessary now that we resolve extensions in the golden files.

Add a patch (and a command to the Makefile to apply the patch) to turn
off the message_set_wire_format which is not available from Go. This
could be enabled with `-tags protolegacy`, but this produces some
differences we don't care to correct.