Upgrade golangci-lint to 1.45 from 1.43
Upgrade golangci-lint to 1.45 as the old version is not working with Go
1.18.

Fix a couple of new lint errors from the upgrade by adding an ignore.
The deprecated method is still fine for the use case, and the
alternative suggested seems rather more involved than a simple
substitution.

Disable the `ireturn` linter - it does not let you return interfaces,
yet the protobuf api is full of interfaces.