Add extension options to all ranges
Apply the options specified on an `extension` clause in a message to all
ranges defined on that line and not just the last one. An `extensions`
clause can specify multiple comma-separated ranges, and a single options
block at the end of the line. That options block should apply to all the
ranges defined, not just the last one.

e.g.

    extensions 100 to 110, 200 to 220, 1000 to max [(opt) = val];

The `(opt) = val` option should apply to all three ranges. Prior to this
change, the following would be parsed as correct whereas protoc rejects
it:

    extensions 100 to 100 [(opt1) = val], 200 to 220 [(opt2) = val];

and it would also accept options on reserved ranges which protoc does
not support.

Fixes: https://github.com/alecthomas/protobuf/issues/36