parser: Fix successive numeric escapes, rune vs byte
Fix unquoting of successive numeric escape sequences so that the second
one straight after the first is actually processed as an escape sequence
and not literal chars.

Limit the length of numeric sequences to 2 or 3 whether hex or octal,
and emit an error if the octal value is too large.

Write numeric escapes as bytes not runes as non-ascii byte values are
interpreted differently as runes and becomes incorrect.