parser: Capture empty ProtoText as non-nil ProtoText
Capture the value of an empty ProtoText element (`{}`) as a non-nil
ProtoText struct that has no contents, rather than as a nil ProtoText
pointer. If it is captured as a nil pointer then the stringer for
ProtoText does not know a value is actually prototext and emits
`UNKNOWN-VALUE` instead.

Handle an empty ProtoText by stringing as `{}` so we round-trip properly
back to string.