compiler: Handle partial extension names in prototext
Handle extension names in prototext when defining an option that are not
full names, but instead are relative to their lexical scope.

When the prototext decoder encounters an extension name (inside square
brackets: `foo { [ext]: value }`) it uses the given
`ExtensionTypeResolver` to look up the name `ext` in this example.
However this is not a full name - the full name of the extension is
`pkg.ext` where `pkg` is the package in which `ext` is defined. It could
also be longer if the extension is defined inside a message.

However, even though the method `FindExtensionByName` takes a
`FullName`, there is no way for the prototext decode to actually know
the full name as it is relative to the outer protobuf definitions, and
none of that is visible to the prototext decoder. This would appear to
be something of a special case - prototext embedded inside protobuf. I
assume that normally prototext has fully-qualified FullNames in the
square brackets for extensions.

Handle this special case by wrapping the resolver with the `types`
registry built during type analysis (which also contains extensions) and
tracking the scopes as we walk the parser AST while we expand options.
This allows us to first look up the full name of an extension given the
scope it is in and then look it up in the resolver.

Add an additional test case specifically for this scenario.

While we're here, remove the tests for uninterpreted options - that
should have been removed before as options are now properly tested with
end-to-end protoc comparisons, and now they interfere with some minor
refactoring needed to make use of `types`.