Search global types before dynamic registry
When looking up the Registry for extensions and messages, first look in
protoreflect.GlobalTypes so that concrete types are used where they are
present in the program binary. This helps when the compiler is used as
library in another program that has some concrete types built in
(i.e. it has linked with generated protobuf code).

This issue has become apparent when working github.com/foxygoat/jig
directly with proto files. The dynamically generated FileDescriptorSet
could not be type asserted against *annotations.HttpRule.

Signed-off-by: Julia Ogris <julia.ogris@gmail.com>